Mintomic for S390 on zOS
This is for 31bit AMODE only, XPLINK linkage.
Extending for 64 bit should be easy based on zLinux
64bit implementation and this.
Inefficient barriers as they are function calls. In
fact whole implementation is function call based.

No idea if mintsystem or mintpack work outside of the
parts necessary for the test suite.

This is based on using xlC as the compiler, and that the
resulting calls will be used in C++. xlc suite doesn't
allow inlined asm in C++, but it could be done for C.

Again, uses older instructions, assuming old hardware could
be the target.

Many of the _gcc files should really be _pthreads, and as zos has
pthreads, I'm using them unadulterated. Unlike the other unices
encountered zos won't cast pthread_t to uint32, so we mangle it to
do so.

Change-Id: Ic486f73d292479bdc7f675e8057dbacf0b361098
Reviewed-on: https://gerrit.teamwpc.local/205
Reviewed-by: Tom Quarendon <tom.quarendon@teamwpc.co.uk>
Reviewed-by: Alan Clucas <alan.clucas@teamwpc.co.uk>
Tested-by: Alan Clucas <alan.clucas@teamwpc.co.uk>